updated with more modern iCloud Photo Library paths

+++ photo-importer.pl
@@ -1,9 +1,7 @@
 #!/usr/bin/env perl
 
 # Written by Breandan Dezendorf <breandan@dezendorf.com>
-# Released under the GPLv2.
-
-# https://github.com/bwdezend/photo-importer
+# Released under the GPL.
 
 use warnings;
 use strict;
@@ -15,6 +13,9 @@ use DBI;
 use DBD::SQLite;
 use Digest::SHA;
 use Getopt::Long;
+use POSIX qw(strftime);
+
+use Time::localtime;
 
 use Image::ExifTool;
 use File::Path;
@@ -25,8 +26,21 @@ my $debug       = undef;
 my $help        = undef;
 my $directory   = undef;
 my $reset       = undef;
-my $storage_dir = "/Volumes/photos/MasterImages";
-my $logfile     = $storage_dir . "/photo-importer.log";
+my $storage_dir = "/mnt/nfs/photos/MasterImages";
+my $logfile     = $storage_dir . "/photostream2lightroom.log";
+my $lockfile    = "/tmp/photo-import.lock";
+my $sleep       = 1;
+my $delete      = undef;
+my $no_import   = undef;
+my $icloud      = 1;
+my $reimport    = undef;
+my $unlink      = undef;
+
+my $year = strftime "%Y", localtime;
+my $month = strftime "%m", localtime;
+my $day = strftime "%d", localtime;
+
+
 
 &Getopt::Long::Configure( 'pass_through', 'no_autoabbrev' );
 &Getopt::Long::GetOptions(
@@ -36,22 +50,25 @@ my $logfile     = $storage_dir . "/photo-importer.log";
     'directory=s' => \$directory,
     'reset'       => \$reset,
     'logfile'     => \$logfile,
+    'sleep=i'     => \$sleep,
+    'delete'      => \$delete,
+    'no_import'   => \$no_import,
+    'icloud'      => \$icloud,
+    'reimport'    => \$reimport,
+    'unlink'      => \$unlink,
 );
 
 if ( $logfile eq "none" ) { $logfile = undef }
 
-# If you want to import from PhotoStream, set $directory to:
-#
-# $ENV{"HOME"} . '/Library/Application Support/iLifeAssetManagement/assets/sub';
-#
-# We're importing from the Dropbox camera uploads by default instead:
-
+if ( !$directory && $icloud ) { $directory = $ENV{"HOME"} . '/Pictures/Photos Library.photoslibrary/Masters/' . $year . '/' . $month . '/'; }
 if ( !$directory ) { $directory = $ENV{"HOME"} . '/Dropbox/Camera Uploads'; }
+
+if ( !$sleep ) { $sleep = '0' }
 
 my $db_file = $ENV{"HOME"} . '/homefolder/etc/processed.sqlite3';
 
 if ($help) {
-    print "photo-importer\n";
+    print "finddupes\n";
     print "  A utility to scan a given directory looking for image files\n";
     print "   and copy new files to a given storage directory. This is done\n";
     print "   using SHA1 hashes of the files. The files are stored based on\n";
@@ -76,12 +93,23 @@ if ($help) {
 &log_message("Starting run");
 $SIG{'INT'} = sub {
     &log_message("Time for an orderly shutdown");
+    if ($debug) { print " Removing lockfile ($lockfile)\n" }
+    system("rm $lockfile");
     exit 0;
 };
 
 my %digests = ();
 
 my $dbh = DBI->connect( "dbi:SQLite:dbname=$db_file", "", "" );
+
+if (-e $lockfile){ die "Could not lock $lockfile\n" } else { system("touch $lockfile") };
+
+unless (-d $storage_dir){
+	print "$storage_dir doesn't exist. Leaving $lockfile in place!\n";
+	system("touch $lockfile");
+	exit 1;
+}
+
 
 if ($reset) {
     $dbh->do("DROP TABLE IF EXISTS hashes");
@@ -105,29 +133,29 @@ if ( $copy_count gt 0 ) {
     &terminal_notify($copy_count);
 }
 
+system("rm $lockfile");
 &log_message("Run finished");
 
 exit 0;
 
 sub wanted {
-    my $digest   = undef;
-    my $sha      = Digest::SHA->new('SHA1');
-    my $exifTool = new Image::ExifTool;
+    my $digest        = undef;
+    my $sha           = Digest::SHA->new('SHA1');
+    my $exifTool      = new Image::ExifTool;
+    my $should_unlink = undef;
 
     if ( -f $_ ) {
         $sha->addfile($_);
         $digest = $sha->hexdigest;
 
         my $basename = basename($File::Find::name);
-        if ( $results->{$digest} ) {
+        if ( !$reimport && $results->{$digest} ) {
             &log_message("exists:  $digest - $basename") if ($debug);
-        }
-        else {
+        } elsif ( $basename =~ m/.xmp/ ) {
+           print "not processing xmp files\n" if ($debug);
+        } else {
             my $dest_dir = undef;
-            &log_message("unique:  $digest - $basename") if ($verbose);
-            $sth =
-              $dbh->prepare("INSERT INTO hashes(Name, Hash) VALUES (?, ?)");
-            $sth->execute( $basename, $digest );
+            &log_message("unique:  $digest - $basename") if ($debug);
 
             my $info = $exifTool->ImageInfo($_);
 
@@ -138,45 +166,90 @@ sub wanted {
                 my $year  = $1;
                 my $date  = $3;
                 my $month = $2;
-                $date =~ s/:/-/g;
-                $dest_dir =
-                    $storage_dir . "/"
-                  . $year . "/"
-                  . $year . "-"
-                  . $month . "/"
-                  . $year . "-"
-                  . $month . "-"
-                  . $date;
-                print "storage_dir: $dest_dir\n" if ($debug);
-                unless ( -d "$dest_dir" ) {
-                    mkpath($dest_dir);
-                    print "  created $dest_dir\n" if ($verbose);
+                if ($year == "0000"){
+                   print "ERROR: invalid exif\n";
+                } else {
+	                $date =~ s/:/-/g;
+	                $dest_dir =
+	                    $storage_dir . "/"
+	                  . $year . "/"
+	                  . $year . "-"
+   	    	          . $month . "/"
+   		              . $year . "-"
+   	        	      . $month . "-"
+                	  . $date;
+                	print "storage_dir: $dest_dir\n" if ($debug);
+                	unless ( -d "$dest_dir" ) {
+                	    mkpath($dest_dir);
+                	    print "  created $dest_dir\n" if ($verbose);
+                	}
+                	unless ($no_import) {
+	                	$sth = $dbh->prepare("INSERT INTO hashes(Name, Hash) VALUES (?, ?)");
+    	            	$sth->execute( $basename, $digest );
+    	            }
                 }
+
             }
             if ($dest_dir) {
                 my $image_file = $dest_dir . '/' . $basename;
                 if ( -e $image_file ) {
+                    print "Calculating SHA for $image_file\n" if ($debug);
                     $sha->addfile($image_file);
                     my $exists = $sha->hexdigest;
 
-                    &log_message("nocopy:  $digest - $basename");
-                    &log_message("nocopy:  $exists - $image_file");
+                    &log_message("nocopy:  $digest - $basename == $image_file");
+
+                 	if ($unlink){
+                 	    if ( $exists eq $digest ) {
+							print "unlink: $basename \n";
+							unlink $basename;
+						} else {
+						    print "unlink failed:\n    $exists\n    $digest\n don't match\n";
+						}
+            		}               	
+
                 }
                 else {
-                    &log_message("copy:    $digest - $basename");
+                    unless ($no_import) {
+	                    &log_message("copy:    $digest - $basename into $image_file");
+	                    my $source_file = $_;
+                    	copy $source_file, $image_file or die $!;
 
-                    #print "copy:    $digest - $basename\n";
-                    copy $_, $image_file or die $!;
-                    $copy_count++;
+                    	if ($unlink){
+		                    print "Calculating SHA for $image_file\n" if ($debug);
+        		            $sha->addfile($image_file);
+		                    my $exists = $sha->hexdigest;
+		                    if ( $exists eq $digest ) {
+								print "unlink: $basename \n";
+								unlink $basename;
+							} else {
+						  		print "unlink failed:\n    $exists\n    $digest\n don't match\n";
+							}
+                		}               	
+
+	                    my $xmp_sidecar_src = $source_file . ".xmp";
+	                    my $xmp_sidecar_dest = $image_file . ".xmp";
+                    	if ( -e $xmp_sidecar_src) {
+                    		&log_message("copy sidecar:  $xmp_sidecar_src - $xmp_sidecar_dest");
+                    		copy $xmp_sidecar_src, $xmp_sidecar_dest or die $!;
+                    		if ($unlink){
+                    		  print "unlink sidecar: $xmp_sidecar_src \n";
+                    		  unlink $xmp_sidecar_src;
+                    		}               	
+                    	}
+	                    $copy_count++;
+                    }
                 }
             }
             else {
-                print
+                unless ($basename =~ m/.xmp/){
+                    print
 "skipping $basename: unable to get EXIF -- DateTimeOriginal\n";
+				}
             }
         }
         $digests{$digest} = $File::Find::name;
-
+        sleep($sleep);
     }
 }
 
@@ -194,14 +267,12 @@ sub log_message {
 
 sub terminal_notify {
     my $count = shift;
-    if ( -x "/usr/local/bin/terminal-notifier2" ) {
+    if ( -x "/usr/local/bin/terminal-notifier" ) {
         my $NOTIFY =
-'/usr/local/bin/terminal-notifier -title "photo-importer" -message "'
+'/usr/local/bin/terminal-notifier -title "photostream2lightroom" -message "'
           . $count
           . ' new messages added to photo vault"';
         my $result = `$NOTIFY`;
     }
 
 }
-
-
